backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
yfinance==0.2.38
edgartools==2.5.0
pandas==2.2.1
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
xgboost==2.0.3
tqdm==4.66.2
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
sqlalchemy==2.0.29
aiosqlite==0.20.0
psycopg2-binary==2.9.11
pytest==8.1.1
httpx==0.27.0
lxml==5.2.1
html5lib==1.1
